app/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
redis
rq
pydantic
langfuse
openai
pyyaml
